polish tipc serving
