+++ deploy/pdserving/web_service_det.py
@@ -18,63 +18,13 @@ import numpy as np
 import cv2
 import base64
 # from paddle_serving_app.reader import OCRReader
-from ocr_reader import OCRReader, DetResizeForTest
+from ocr_reader import OCRReader, DetResizeForTest, ArgsParser
 from paddle_serving_app.reader import Sequential, ResizeByFactor
 from paddle_serving_app.reader import Div, Normalize, Transpose
 from paddle_serving_app.reader import DBPostProcess, FilterBoxes, GetRotateCropImage, SortedBoxes
 
-import yaml
-from argparse import ArgumentParser,RawDescriptionHelpFormatter
-
 _LOGGER = logging.getLogger()
 
-class ArgsParser(ArgumentParser):
-    def __init__(self):
-        super(ArgsParser, self).__init__(
-            formatter_class=RawDescriptionHelpFormatter)
-        self.add_argument("-c", "--config", help="configuration file to use")
-        self.add_argument(
-            "-o", "--opt", nargs='+', help="set configuration options")
-
-    def parse_args(self, argv=None):
-        args = super(ArgsParser, self).parse_args(argv)
-        assert args.config is not None, \
-            "Please specify --config=configure_file_path."
-        args.conf_dict = self._parse_opt(args.opt, args.config)
-        return args
-
-    def _parse_helper(self, v):
-        if v.isnumeric():
-            if "." in v:
-                v = float(v)
-            else:
-                v = int(v)
-        elif v == "True" or v == "False":
-            v = (v == "True")
-        return v
-
-    def _parse_opt(self, opts, conf_path):
-        f = open(conf_path)
-        config = yaml.load(f, Loader=yaml.Loader)
-        if not opts:
-            return config
-        for s in opts:
-            s = s.strip()
-            k, v = s.split('=')
-            v = self._parse_helper(v)
-            print(k,v, type(v))
-            cur = config
-            parent = cur
-            for kk in k.split("."):
-                if kk not in cur:
-                     cur[kk] = {}
-                     parent = cur
-                     cur = cur[kk]
-                else:
-                     parent = cur
-                     cur = cur[kk]
-            parent[k.split(".")[-1]] = v
-        return config
 
 class DetOp(Op):
     def init_op(self):
