+++ test_tipc/configs/ch_ppocr_server_v2.0_det/model_linux_gpu_normal_normal_serving_python_linux_gpu_cpu.txt
@@ -9,10 +9,10 @@ trans_model:-m paddle_serving_client.convert
 --serving_client:./deploy/pdserving/ppocr_det_server_2.0_client/
 serving_dir:./deploy/pdserving
 web_service:web_service_det.py --config=config.yml --opt op.det.concurrency=1
-op.det.local_service_conf.devices:null|0
+op.det.local_service_conf.devices:"0"|null
 op.det.local_service_conf.use_mkldnn:True|False
 op.det.local_service_conf.thread_num:1|6
 op.det.local_service_conf.use_trt:False|True
 op.det.local_service_conf.precision:fp32|fp16|int8
 pipline:pipeline_rpc_client.py|pipeline_http_client.py
---image_dir:../../doc/imgs_words_en
+--image_dir:../../doc/imgs
