+++ test_tipc/configs/ch_ppocr_server_v2.0_rec/model_linux_gpu_normal_normal_serving_python_linux_gpu_cpu.txt
@@ -1,6 +1,6 @@
 ===========================serving_params===========================
 model_name:ocr_rec_server
-python:python3.7
+python:python3.7|cpp
 trans_model:-m paddle_serving_client.convert
 --dirname:./inference/ch_ppocr_server_v2.0_rec_infer/
 --model_filename:inference.pdmodel
@@ -9,7 +9,7 @@ trans_model:-m paddle_serving_client.convert
 --serving_client:./deploy/pdserving/ppocr_rec_server_2.0_client/
 serving_dir:./deploy/pdserving
 web_service:web_service_rec.py --config=config.yml --opt op.rec.concurrency=1
-op.rec.local_service_conf.devices:null|0
+op.rec.local_service_conf.devices:"0"|null
 op.rec.local_service_conf.use_mkldnn:True|False
 op.rec.local_service_conf.thread_num:1|6
 op.rec.local_service_conf.use_trt:False|True
